requirements.txt
# Core dependencies
streamlit==1.45.1
torch==2.7.0
torchvision==0.22.0
transformers==4.52.3

# AI/ML libraries
accelerate==1.7.0
sentence-transformers==4.1.0
huggingface-hub==0.32.1
safetensors==0.5.3

# Computer Vision
opencv-python-headless==4.11.0.86
pillow==11.2.1

# Data processing
numpy==2.2.6
pandas==2.2.3
matplotlib==3.10.3

# CLIP
git+https://github.com/openai/CLIP.git

# Streamlit theme
streamlit_theme==0.58.0

# Other dependencies
aiohappyeyeballs==2.6.1
aiohttp==3.12.1
aiosignal==1.3.2
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
dill==0.3.8
filelock==3.18.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
ImageHash==4.3.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
narwhals==1.41.0
networkx==3.4.2
nltk==3.9.1
packaging==24.2
peft==0.15.2
pipreqs==0.4.13
plotly==6.1.1
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
pyarrow==20.0.0
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.25.1
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.8.0
six==1.17.0
smmap==5.0.2
st-theme==1.2.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.13
xxhash==3.5.0
yarg==0.1.10
yarl==1.20.0
